+++ changelog.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Updated GitHub Actions dispatch for homebrew
+
 ## [1.33.0] - 2026-08-20
 
 ### Added